+++ app/lib/data.ts
@@ -178,26 +178,3 @@ export async function fetchTwoteeksResult() {
         throw new Error('Failed to fetch two weeks result.');
     }
 }
-
-
-export async function addVoteEveryTwoWeeks() {
-    try {
-        await sql`
-            UPDATE drink
-            SET totalvoted = totalvoted + voted
-        `;
-    } catch (error) {
-        return { message: 'Database Error: Failed to add vote every two weeks' };
-    }
-}
-
-export async function deleteVoteEveryTwoWeeks() {
-    try {
-        await sql`
-            UPDATE drink
-            SET voted = 0
-        `;
-    } catch (error) {
-        return { message: 'Database Error: Failed to delete vote every two weeks' };
-    }
-}
